Add Comment collection components
Adds Comments and Comment comment for showing comment data on the UpdateTaskForm modal.
The changes includes adjusting of comments index in CommentsController to eager load
associated users with their avatar_image url as well as including the total comments count
of the selected task. The changes also includes renaming of css stylesheet to be used by all
comment components.