Separate tasks from boards index API
The changes mainly separates the Task from the boards index
for more readability and much cleaner code on boards_controller. The tasks
query is instead placed on the index of tasks_controller, whereas the previouse query for tasks summary is moved to 'summary' endpoint. The changes also includes adding more test cases on the boards_spec and tasks_spec request tests.