create new form component for creating tasks
The commit creates new separate components used for the form creating
tasks. The components is consists of the main component, which is the
Form, and the two child components, which are the LeftPanel and RightPanel
components.

The change makes the code more readable and much easier to debug or add new
features on the form.`